FIR checker: relocate sealed class ctor call checker
